refactor(viewer): extract viewer_watchlist_core.jsx from viewer.jsx
Pure file split, no behavior change. Step 2 of mp-pxxc.
